feat: implement Claude usage provider and integrate into dashboard
